+++ agential/cog/self_refine/functional.py
@@ -59,7 +59,13 @@ def _prompt_agent(
         prompt=prompt,
         additional_keys=additional_keys,
     )
+    print("<PROMPT AGENT=====================================================>")
+    print(prompt)
+    print("<PROMPT AGENT=====================================================>")
     out = llm(prompt)
+    print("<OUT AGENT=====================================================>")
+    print(repr(out.output_text))
+    print("<OUT AGENT=====================================================>")
     return out
 
 
@@ -125,7 +131,14 @@ def _prompt_critique(
         prompt=prompt,
         additional_keys=additional_keys,
     )
+    print("<PROMPT CRITIC=====================================================>")
+    print(prompt)
+    print("<PROMPT CRITIC=====================================================>")
     out = llm(prompt)
+    print("<OUT CRITIC=====================================================>")
+    print(repr(out.output_text))
+    print("<OUT CRITIC=====================================================>")
+
     return out
 
 
@@ -193,7 +206,14 @@ def _prompt_refine(
         prompt=prompt,
         additional_keys=additional_keys,
     )
+    print("<PROMPT REFINE=====================================================>")
+    print(prompt)
+    print("<PROMPT REFINE=====================================================>")
     out = llm(prompt)
+    print("<OUT REFINE=====================================================>")
+    print(repr(out.output_text))
+    print("<OUT REFINE=====================================================>")
+
     return out
